sync command wip
