+++ Sources/App/routes.swift
@@ -5,8 +5,6 @@ func routes(_ app: Application) throws {
     
     try app.register(collection: ForecastapiController())
     
-    try app.register(collection: SyncController())
-    
     try app.register(collection: S3DataController())
 }
